requirements.txt
numpy>=1.17.4
pandas>=0.25.3
xlrd>=1.2.0
matplotlib>=3.1.1
joblib>=0.15.1
tensorflow==2.2.0
sklearn
scikit-learn>=0.23.1
grpcio==1.29.0
grpcio-tools==1.29.0
grpcio-status==1.29.0
future==0.16.0
observable>=1.0.3
astor==0.8.1
pyinstaller==3.6
